requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
currencyapicom==0.1.1
everapi==0.1.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
Markdown==3.5.2
oauthlib==3.2.2
pydantic==2.6.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
python-dotenv==1.0.1
python-telegram-bot==20.8
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.2.0
